requirements.txt
groq
python-dotenv
speechrecognition
pydub
pyaudio
gtts
elevenlabs
gradio